+++ src/main/java/conspire/relics/AbstractConspireRelic.java
@@ -9,7 +9,7 @@ public abstract class AbstractConspireRelic extends AbstractRelic {
     public AbstractConspireRelic(String id, RelicTier tier, LandingSound sfx) {
         super(id, "", tier, sfx);
         img = ImageMaster.loadImage(Conspire.relicImage(id));
-        largeImg = ImageMaster.loadImage(Conspire.relicLargeImage(id));
         outlineImg = ImageMaster.loadImage(Conspire.relicOutlineImage(id));
+        this.imgUrl = Conspire.relicImgUrl(id);
     }
 }